requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
curl_cffi==0.10.0
distro==1.9.0
frozendict==2.4.6
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
mcp==1.7.1
multitasking==0.0.11
numpy
openai==1.77.0
openai-agents @ git+https://github.com/openai/openai-agents-python.git@f9763495b86afcf0c421451a92200e1141fa8dcb
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.7
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.4
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
yfinance==0.2.58
matplotlib
scikit-learn
